Make filename length configurable
Improve random filename generation by allowing users to configure how
long the filename should be, and checking, when possible, whether a file
with the generated name already exists before trying to upload an image
with that name. This prevents the uploader from accidentally overwriting
an existing file on several server types, such as WebDAV servers.